backend/requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
requests==2.31.0
python-slugify==8.0.1
tmdbv3api==1.9.0
